refactor: rename api to extension.api
